@@ -3,35 +3,41 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/data-grid": "^4.0.0-alpha.37",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@stripe/react-stripe-js": "^1.4.1",
-    "@stripe/stripe-js": "^1.17.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "chart.js": "^3.5.1",
-    "country-state-city": "^3.0.1",
-    "overlay-navbar": "^1.0.4",
-    "react": "^17.0.2",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.7",
+    "@mui/lab": "^5.0.0-alpha.162",
+    "@mui/material": "^5.15.7",
+    "@mui/styles": "^5.15.7",
+    "@mui/x-data-grid": "^6.19.3",
+    "@redux-devtools/extension": "^3.3.0",
+    "@stripe/react-stripe-js": "^2.4.0",
+    "@stripe/stripe-js": "^2.4.0",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "ajv": "^8.17.1",
+    "ajv-keywords": "^5.1.0",
+    "axios": "^1.6.7",
+    "chart.js": "^4.4.1",
+    "country-state-city": "^3.2.1",
+    "overlay-navbar": "^1.2.3",
+    "react": "^18.2.0",
     "react-alert": "^7.0.3",
     "react-alert-template-basic": "^1.0.2",
-    "react-chartjs-2": "^3.0.4",
-    "react-dom": "^17.0.2",
+    "react-chartjs-2": "^5.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-icons": "^4.2.0",
+    "react-icons": "^5.0.1",
     "react-js-pagination": "^3.0.3",
-    "react-material-ui-carousel": "^2.3.1",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.1",
+    "react-material-ui-carousel": "^3.4.2",
+    "react-redux": "^9.1.0",
+    "react-router-dom": "^6.22.0",
+    "react-scripts": "5.0.1",
+    "redux": "^5.0.1",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0",
-    "web-vitals": "^1.1.2",
+    "redux-thunk": "^3.1.0",
+    "web-vitals": "^3.5.2",
     "webfontloader": "^1.6.28"
   },
   "scripts": {
@@ -58,5 +64,5 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://192.168.0.104:8080"
+  "proxy": "http://localhost:4000"
 }
