@@ -6,9 +6,9 @@ import { clearErrors, updatePassword } from "../../actions/userAction";
 import { useAlert } from "react-alert";
 import { UPDATE_PASSWORD_RESET } from "../../constants/userConstants";
 import MetaData from "../layout/MetaData";
-import LockOpenIcon from  "@material-ui/icons/LockOpen"
-import LockIcon from  "@material-ui/icons/Lock"
-import VpnKeyIcon from "@material-ui/icons/VpnKey"
+import LockOpenIcon from "@mui/icons-material/LockOpen"
+import LockIcon from "@mui/icons-material/Lock"
+import VpnKeyIcon from "@mui/icons-material/VpnKey"
 
 const UpdatePassword = ({history}) => {
   const dispatch = useDispatch();
