@@ -1,6 +1,6 @@
 import React from "react";
 import "./Contact.css";
-import { Button } from "@material-ui/core";
+import { Button } from "@mui/material";
 
 const Contact = () => {
   return (
