@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import {Provider} from "react-redux"
 import store from "./store"
@@ -10,11 +10,14 @@ const options={
   positions:positions.BOTTOM_CENTER,
   transitions:transitions.SCALE,
 }
-ReactDOM.render(
+
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <Provider store={store}>
     <AlertProvider template={AlertTemplate}{...options}>
     <App />
     </AlertProvider>
-  </Provider>,
-  document.getElementById('root')
+  </Provider>
 );
