@@ -109,7 +109,7 @@ exports.deleteOrder = catchAsyncErrors(async (req, res, next) => {
     if (!order) {
         return next(new ErrorHandler("order not found with this id", 404))
     }
-    await order.remove()
+    await order.deleteOne()
     res.status(200).json({
         success: true,
     });
