@@ -1,18 +1,19 @@
-const nodeMailer = require("nodemailer")
+const { Resend } = require('resend');
+
 const sendEmail = async (options) => {
-    const transporter = nodeMailer.createTransport({
-        service: process.env.SMPT_SERVICE,
-        auth: {
-            user: process.env.SMPT_MAIL,
-            pass: process.env.SMPT_PASSWORD,
-        }
-    })
-    const mailOptions = {
-        from: process.env.SMPT_MAIL,
-        to: options.email,
-        subject: options.subject,
-        text: options.message,
+    const resend = new Resend(process.env.RESEND_API_KEY);
+
+    try {
+        await resend.emails.send({
+            from: 'onboarding@resend.dev', // Use default or your verified domain
+            to: options.email,
+            subject: options.subject,
+            text: options.message,
+        });
+    } catch (error) {
+        console.error("Email sending failed:", error);
+        throw new Error("Email sending failed");
     }
-    await transporter.sendMail(mailOptions);
 }
+
 module.exports = sendEmail;
