@@ -1,31 +1,23 @@
-import React, { Fragment } from 'react'
-import { useSelector } from 'react-redux'
-import { Redirect, Route } from 'react-router-dom'
+import React from 'react';
+import { useSelector } from 'react-redux';
+import { Navigate, Outlet } from 'react-router-dom';
 
-const ProtectedRoute = ({isAdmin,component:Component,...rest}) => {
-    const {loading,isAuthenticated,user} = useSelector((state)=>state.user)
-    
-    return (
-        <Fragment>
-            {loading===false &&(
-               <Route 
-               {...rest} 
-               render={
-                 (props)=>{
-                     if (isAuthenticated=== false) {
-                         return <Redirect to="/login"/>
-                        }
-                        if (isAdmin===true &&user.role!=="admin") {
-                         return <Redirect to="/login"/>
-                        }
-                     return <Component {...props}/>
-                 }  
-               }/> 
-            )}
-        </Fragment>
-    )
+const ProtectedRoute = ({ isAdmin, children }) => {
+    const { loading, isAuthenticated, user } = useSelector((state) => state.user);
 
-}
+    if (loading) {
+        return null; // or a loader/spinner
+    }
 
-export default ProtectedRoute
- 
+    if (isAuthenticated === false) {
+        return <Navigate to="/login" />;
+    }
+
+    if (isAdmin === true && user.role !== "admin") {
+        return <Navigate to="/login" />;
+    }
+
+    return children ? children : <Outlet />;
+};
+
+export default ProtectedRoute;
