@@ -3,10 +3,10 @@ import {
     combineReducers,
     applyMiddleware
 } from "redux";
-import thunk from "redux-thunk";
+import { thunk } from "redux-thunk";
 import {
     composeWithDevTools
-} from "redux-devtools-extension";
+} from "@redux-devtools/extension";
 import {
     newProductReducer,
     newReviewReducer,
